+++ pkg/gatewayserver/gatewayserver_test.go
@@ -716,7 +716,7 @@ func TestGatewayServer(t *testing.T) {
 									},
 								},
 								ExpectLocation: ttnpb.Location{
-									Source: ttnpb.SOURCE_REGISTRY,
+									Source: ttnpb.SOURCE_GPS,
 								},
 							},
 							{
@@ -728,7 +728,7 @@ func TestGatewayServer(t *testing.T) {
 									},
 								},
 								ExpectLocation: ttnpb.Location{
-									Source: ttnpb.SOURCE_REGISTRY,
+									Source: ttnpb.SOURCE_GPS,
 								},
 							},
 							{
@@ -763,7 +763,7 @@ func TestGatewayServer(t *testing.T) {
 								})
 								a.So(err, should.BeNil)
 								gtw.Antennas[0].Location = ttnpb.Location{
-									Source: ttnpb.SOURCE_REGISTRY,
+									Source: ttnpb.SOURCE_GPS,
 								}
 								gtw.UpdateLocationFromStatus = tc.UpdateLocation
 								gtw, err = is.Get(ctx, &ttnpb.GetGatewayRequest{